Implementing the code for editing the product

+++ src/pages/admin/dashboard/product/Product.jsx
@@ -1,10 +1,12 @@
 import React, { useEffect, useState } from 'react'
 import { handleDeleteOperation, handleGetOperation } from '../../../../handleOperation/handleOperation'
 import { apiLinks } from '../../../../handleOperation/apiLinks'
+// import { useNavigate } from 'react-router-dom'
 
 const Product = () => {
 
   const [formData,setFormData] = useState([])
+  const [count , setCount] = useState(1)
 
 
   useEffect(()=>{
@@ -20,21 +22,26 @@ const Product = () => {
           setFormData(result.data.data)
         }
       } catch (error) {
-        console.error(error);
-        alert("Failed to fetch product")
+        console.error("Failed to fetch product:", error);
+        alert("Failed to fetch product");
       }
 
     }
     fetchProduct()
-  },[])
+  },[count])
 
 
   const handledDelete = async(id)=>{
-    alert(id)
-    const result = await handleDeleteOperation(apiLinks.deleteProduct(id))
+    
+    const isDelete = confirm("Do you want to delete ")
+    if(isDelete){
+      const result = await handleDeleteOperation(apiLinks.deleteProduct(id))
 
-    console.log(result);
-
+      if(result.status == 200){
+        setCount((prev)=>prev+1)
+        // location.href = '/admin/product'
+      }
+    }
   }
 
   return (
